[Refactor] refactor caching iceberg metadata (#44218)

+++ fe/fe-core/src/main/java/com/starrocks/common/Config.java
@@ -2123,7 +2123,7 @@ public class Config extends ConfigBase {
      * size of iceberg worker pool
      */
     @ConfField(mutable = true)
-    public static long iceberg_worker_num_threads = Runtime.getRuntime().availableProcessors();
+    public static int iceberg_worker_num_threads = Runtime.getRuntime().availableProcessors();
 
     /**
      * size of iceberg table refresh pool
